\encoding{utf8}
\name{Density.T.HoldOut}
\alias{Density.T.HoldOut}
\docType{package}
\title{
Density.T.HoldOut: Non-combinatorial T-estimation Hold-Out for density estimation
}
\description{
Implementation in the density framework of the non-combinatorial algorithm and its greedy version, introduced by Magalhães and Rozenholc (2014), for T-estimation Hold-Out proposed in Birgé (2006, Section 9). The package provide an implementation which uses several families of estimators (regular and irregular histograms, kernel estimators) which may be used alone or combined. As a complement, provides also a comparison with other Held-Out derived from least-squares and maximum-likelihood. 
}
\details{
\tabular{ll}{
Package: \tab Density.T.HoldOut\cr
Type: \tab Package\cr
Version: \tab 1.00\cr
Date: \tab 2014-01-08\cr
License: \tab GPL(>=2)\cr
}
}
\author{
Nelo Magalhães and Yves Rozenholc.

Maintainer: Nelo Magalhães <nelo.moltermagalhaes@gmail.com>

}
\references{
L. Birgé, "Model selection via testing: an alternative to (penalized) maximum likelihood estimators.", Ann. Institut Henri Poincaré Probab. et Statist., 42, 273--325, (2006)
    
N. Magalhães and Y. Rozenholc, "A non-combinatorial algorithm for T-estimation.", (2014)
}

